swap over to the new lint command.
all new!
